fix: Ревью #3137: осмотр пьяных перенесён в Shared, текст без задержки пинга
